Add waseem clustering analysis
Just to check out if the clustering gives us more significant results in the waseem dataset.
However, by looking at the plots, the results seem very similar.
We can also notice that there are less and more specific topics in the waseem dataset.
This is in line with their work since they indicates that they searches on very specific hashtags (for example the hashtag of a TV show)